libbladeRF: Cache state of attached expansion board
We can read and cache this state when we open the device,
and update this cached value on a change.

This is done to remove a superfluous config read from the FPGA
during the frequency retuning operations. (At this point, this does
make a difference when shooting for sub-millisecond tuning.)