hdl: Ported Trigger IP for use in bladeRF project
The trigger has been configured to be active-low to allow for
effective use of the FPGA internal pull-ups in systems where some
trigger-slaves might come up when there is no master yet driving the bus
to a defined state.

If no external pull-resistors are in place, the bus state would
essentially be random. By using an active-low trigger and eventually
enabling the weak pull-ups on mini_exp1, the trigger line will always be
in a defined state whether or not a master or external circuitry is
present.