libbladeRF: Fix #404, XB-200 filter bank selection regression in 41476cb
The fix for issue #394 introduced another defect via:

    /* Update state information regarding automatically selected filter.
     * Invalidate the entry if we're not performing automatic selection. */
    if (filter == BLADERF_XB200_AUTO_1DB || filter == BLADERF_XB200_AUTO_3DB) {
        dev->auto_filter[module] = filter;
    } else {
        dev->auto_filter[module] = -1;
    }

The problem is that xb200_auto_filter_selection() ends up calling this with
filter set to the desired filterbank. We should only apply this state update
when getting called from bladerf_xb200_set_filterbank(), and not when we're
updating the filter selection from xb200_auto_filter_selection().