libbladeRF: Prepend installation prefix to file search paths
When searching for FPGA bitstreams or other files, we should be
searching locations relative to the installation prefix.

In an effort to avoid breaking existing packaging scripts, the
previously used paths "/usr/share/Nuand/bladeRF" and
"/etc/Nuand/bladeRF" have been kept in the search list, but at a lower
priority than the items with the installation prefix.