libbladeRF: End bursts with 3 zero-samples, not 2.
See "Figure 9: TX data interface" in the LMS6002D data sheet for the
register stages that drive this requirement.

Due to a extra clock cycle delay in bladeRF FPGA <= v0.3.3, we were
"getting away" with a 2-sample 0-pad. However, FPGA v0.3.4 removes the
extraneous delay, which requires us to end the burst properly, with 3
zero samples.