lms: Added support for multi-write of TX and RX NINT/NFRAC registers.
Achieved by adding extra address bit (now 8 bits) to Avalon-MM
interface. The MSB indicates a multi-write instruction. For TX, the
addresses are 0x90-0x93. For RX, the addresses are 0xA0-0xA3. The
individual writes are queued up until a write to the final address is
detected (0x93 or 0xA3), then all four PLL configuration registers are
written to the LMS chip in one contiguous chip select burst.

Testbench updated to be self-checking.