Simply xb200 filter configuration routines.
Handling of the XB-200 hard and auto filters is seperated into two
paths to avoid the confusion introduced in the fix for #404.

xb200_set_filterbank is the internal API entry point for the external
bladerf_xb200_set_filterbank. If the configured filter is a hard
filter, then internal state is updated and set_filterbank_mux is
invoked to realize the hardware changes.

If the configured filter is an automatic filter, then internal state
is updated and xb200_auto_filter_selection() is called with the
currently tuned frequency for specificed module.

xb200_auto_filter_selection() then invokes set_filterbank_mux to
realize the hardware changes.