quartus: Enable/Disable TalkBack.
Enable/Disable the TalkBack feature of Quartus based on SignalTap
insertion and if the user really wants to force the script to enable
and disable it at the end of building.

If TalkBack is disabled, and SignalTap is requested to be added to
the project, then the build will now fail instead of warn about
TalkBack not being enabled.