libbladeRF: Adding ability to set SMB frequency.
Adding the ability to set the SMB frequency when in MIMO master mode.
Also fix a bug where attaching the xb200 and setting MIMO master
would turn off the other clock not allowing an xb200 enabled device
to ever be a master clock supplier.

Being able to set the master frequency is useful when you want to
lock test equipment to the bladeRF oscillator.