hdl: Fixed PPS count interrupts
Interrupts used to fire every second if any of the three counters were off
target. Now, an interrupt will fire only when a new valid count appears
and the associated count value is off target. For example, if the 1 second
count is within the acceptable range, but the 10 second count is not, the
interrupt will only fire at 10 second intervals when the 10 second count
is updated.

Interrupts can only be cleared by the host processor by clearing the bit
specified in the memory map.

Error tolerance is now a 64-bit signed number as opposed to a 'natural',
which is limited to 32 bits.