nios: Working baseline for VCTCXO tuning via PPS signal
Changed ppscal functions to use signed 64-bit integers (int64_t vs. uint64_t)
to be able to discern which direction the trim DAC needs to be adjusted.

Added a global variable to cache the trim DAC setting. This variable shall
only be updated during a call to 'devices.c:vctcxo_trim_dac_write()' and
is used by the VCTCXO tuning alorithm to make adjustments to the trim DAC
without having to read from the DAC.

Fixed bug in ppscal_read() that would only read the 1 second counter
regardless of the address passed into the function.

VCTCXO tuning algorithm in the main program loop actually does something
useful and is able to (over a very long period of time) tune the VCTCXO
to within 10 PPB. A proper PID-type algorithm should be implemented.