nios: Implemented coarse and fine tune algorithm for VCTCXO tuning from 1PPS
Coarse algorithm rails the trim DAC to the min and max voltage supported by
the VCTCXO and creates a line based on the two points (x,y)=(error,DAC_count).
The slope is counts/Hz and the y-intercept is the DAC count with a theoretical
error of 0 Hz. After tuning the the y-intercept, a fine tuning method is used
in which the error is multiplied by a scaled version of the slope based on
the precision of the error (1s, 10s, 100s).

hdl: Changed the 10s error tolerance from 3 to 4.