hdl: Fixed tune reference MUX select inputs and
The tune reference MUX select inputs were connected to a version of the
tune mode signal synchronized to the tune_ref clock domain. This was an
oversight that would cause a 2-3 second delay after resetting the
counters when tuning against a 1PPS reference (because the synchronizer
block contains two flops clocked at 1 Hz). This was corrected and the
MUX now uses the system-synchronous tune_mode signal from the control
register in the Avalon-MM interface.

Renamed the tune_mode signals to include the clock domain (prefixed with
tune_ref_*).

Added ability for the Avalon-MM master to read back the control register
contents.

Updated tcxopps.stp SignalTap file with new signal names.