adsb: Adding FPGA based ADSB decoding.
This commit adds all the files required for using the FPGA on
the bladeRF as an ADSB message decoder.  The decoder uses a low
IF approach for decoding the PPM messages from ADSB.  The FPGA
processing expects a samplerate of 16 MHz and the IF to be at
4 MHz.  The FPGA will mix, filter and turn the result into a
power value.  This power value is then compared against a noise
floor, the preamble is searched and, if found, one of a number
of message decoders are used to decode the data.  If the CRC
does not pass, a bit flipping technique is used to flip the
lowest 5 soft decisions and re-try the CRC.  If successful,
the message is sent into a holding area and into a FIFO which
collects the ADSB messages.  After a timeout period, if messages
have not filled the FIFO, zeros are stuffed until the FIFO is
full enough for a DMA transfer to the host.

An example of a host application for receiving the ADSB messages
is located in the example directory.