build: Fix SignalTap in scripted builds
Simply enabling SignalTap and specifying the .stp file in the project
settings is not sufficient for getting SignalTap into the build. The
.stp file must be translated into project assignments that get appended
to the .qsf file. This is done by calling 'quartus_stp' with the project
name, revision, the .stp file, and the parameter to enable SignalTap in
the project. Quartus_stp must run successfully before Analysis &
Synthesis.

Also added an exception to let non-Web Editions (e.g. Standard and Pro)
of the Quartus tools to add SignalTap files without enabling TalkBack.