fx3_firmware: Use CMake for project configuration and Makefile generation
The combination of CMake and a custom Makefile/toolchain.mk was messy.

This change is to achieve better consistency with the CMake-oriented
workflow of the remainder of the project software and to make it easier
to maintain and change in the future. (CMake options may be used to
enable/disable features.)