libbladeRF: Added high-level trigger implementation
This commit introduces high-level triggering API functions. This allows
multiple devices to be configured to synchornously start RX/TX
operations based upon the assertion of an external I/O signal.

This higher level API is intended to decouple usage from the underlying
register interface and provide some provisions for future
development of other trigger signals (both as official and user
customizations HW/SW customizations).

The general usage of this API is as follows:

    1. Create a bladerf_trigger structure for each trigger in the entire
        system. bladerf_trigger_init() may be used to initialize this
        structure.

    2. Arm 1 trigger master device via bladerf_trigger_arm().

    3. Arm trigger slaves via bladerf_trigger_arm().

    4. Configure and start sample streams. No samples will be
        transmitted or received until the trigger fires. Ensure
        the timeouts used are sufficiently large.

    5. Fire the trigger on the master device via bladerf_trigger_fire().

    6. Transmit and receive samples.

    7. Re-arm triggers via bladerf_trigger_arm() and repeat 5-6 as
        needed.

    8. Disable triggering by setting each trigger's role to
        BLADERF_TRIGGER_ROLE_DISABLED and then calling
        bladerf_trigger_arm(dev, false);