bladeRF-cli: Disarm triggers when shutting down streams
If streams are blocked on triggers, attempts to close the CLI or
stop a stream will "stall" for the duration of the stream timeout.

Generally, timeouts will be set to long values when working with
triggers, yielding a particularly nasty user experience. Therefore,
attempts to disarm triggers are when asked shut down a stream.