libbladeRF: Initialize device sooner.
Initialize the device upon opening it before trying to cancel
any pending retune operations.  In the cases of autoloaded FPGA
images, the sample clocks are not running and can cause the NIOS
to hang when trying to cancel pending retune operations.  This is
alleviated by initializing the device first, ensuring clocks are
running, and then proceeding.