ip: adi: Use .qip in axi_ad9361_hw.tcl fileset
When Qsys generates a system, it copies all the fileset files into its
output directory. This commit changes the file set to be a single .qip
so that only the .qip is copied and the HDL files are referenced from
their original location(s) making modification and debug substantially
easier.