ip: adi: Fix PLL resets
The separate resets for the RX PLL and TX PLL prevented the PLLs from
being merged and thus prevented fitting in the device due to lack of
valid PLL locations. Because we want them to be merged, just use one
reset block to reset "both" PLLs.

In the future (after the TX/RX interfaces have been validated), we can
configure the LVDS cores to have an external PLL to make it explicit.