bladerf-micro: hdl: Rename nios_gpio.spi_mux to nios_gpio.adf_chip_enable
ADF chip enable is a more appropriate name for the signal. The use as a
"SPI mux" is secondary and is intended to block SPI accesses of the
ADF400x or the AD56x1 when the 'other' chip is setting the VCTCXO
control voltage. The idea is to prevent contention on the control
voltage pin if both ADF400x and AD56x1 are enabled at the same time.

The software will have to remember to tristate the AD56x1 DAC before
enabling the ADF400x, but this has always been the case.