hdl/fx3_gpif: consider RX FIFO usage when arbitrating TX/RX
In the event that we are doing a lot of full-duplex operation,
and the FX3 is requesting both RX and TX simultaneously, we
choose TX.  (The reason is probably arbitrary, but I retcon'd
an explanatory comment...)

However, if the RX FIFO is about to burst, we'll pick RX to
give it a chance to drain a bit.

I'm not sure this can happen in reality, but it happens in the
testbench.