hdl: add fabric register probe, integrate with NIOS
This commit adds a "fabric_register_probe" component that can be used
with custom FPGA designs to export some FPGA logic state to the PC
userland.
This is useful for PC applications that want to monitor states, such as
error counters, lock detects in receivers etc.
This commits adds the VHDL implementation and a Qsys component, which
is added to the NIOS-System.

Two instance variables are added to the NIOS-System, which control
whether the registers are exposed into the FPGA design or not. Per
default, these are set to 0, which is also reflected in
build_bladerf.sh. Custom designs can set these to 1 there to export
outputs and/or inputs, respectively.

NIOS2 C code is added to support this component using a 8x32 packet.