host: bladerf2: init ina219 before ad9361
This is a workaround for persistent USB timeouts after AD9361
init when CMAKE_BUILD_TYPE=Release.

If DEBUG is defined, ad9361_verify_fir_filter_coef reads in all
the FIR filter coeffs from the 9361 and makes sure they are as
expected.  If it is NOT defined, it skips this check.

For reasons unknown, if ina219_init is called immediately after
ad9361_ensm_restore_prev_state (the last thing that happens before
ad9361_verify_fir_filter_coef in ad9361_load_fir_filter_coef),
libusb times out and everything goes terribly wrong.  Adding a
usleep(100) is sufficient to avoid this.

There does not seem to be a reason why this should make any
difference whatsoever, but this seems to do the trick.