host: bladeRF-cli: skip init if loading FPGA
If the -l option is specified to bladeRF-cli, assume there is no
FPGA image loaded and skip the initialization part of bladerf_open.

This literally cuts bladeRF-cli -l time in half (4.8 s to 2.4 s)
because of the lengthy AD9361 init.  Additionally, it provides an
escape from situations where the FPGA needs to be reloaded to get
out of a bad state.