ip: opencores: Fix I2C ARST_LVL
The OpenCores I2C ARST_LVL was specified as STD_LOGIC in the _hw.tcl
file. But becuase the top-level is Verilog, this causes Qsys to output
a syntactically incorrect value for the parameter. Solved this by
changing the parameter to an integer, limiting the allowed values to 0:1
and only using the LSB in the Verilog HDL.