hdl: add DC LUT mux correction
Allow the AGC to choose the appropriate DC offset LUT setting for correcting
the DC offset. Essentially, libbladeRF loads the DC offset lookup table from a
previously computed DC calibration table, and NIOS exposes the correction for
each of the gains. The AGC the selects which correction it wants applied based
on its gain setting.