host: libbladeRF: force AD9361 freq change on init
The EVM coming out of the AD9361 is about 1% worse than it should be,
when doing a TX directly after initialization.  However, by changing the
frequency and changing it back, the EVM returns to the expected level.

I suspect libad9361 is missing a calibration step somewhere in the
initialization, but we can work around this by doing two frequency
changes during initialization.