hdl: bladeRF-all: Change direction of Nios control PIO
Changed from "output" to "inout" to accept inputs. In this mode, the
GPIO has two independent buses, one for input, one for output. In order
to maintain previous behavior, which allowed the Nios to read the
current values being output, we must explicitly feed back the output bus
to the input bus.