fx3: skip EP Config during FPGA config if no USB
If the USB interface is not connected, CyU3PUsbGetSpeed() returns
CY_U3P_NOT_CONNECTED.  Previously we treated this as an error, which
caused failures when loading an FPGA bitstream from flash in a headless
configuration.

We now assume a reasonable DMA size and skip the USB configuration
altogether if there is no USB connection.