host: fix builds on 32-bit platforms (#573)
Builds were failing in various ways on 32-bit platforms, due to various assumptions (and the fact that pretty much everything is 64-bit nowadays). This PR should fix all of them so that it builds clean.

* host: libbladeRF/thirdparty/adi: fix build error on 32-bit systems

sizeof(void*) is not necessarily equal to sizeof(uint64_t)

* host: libbladeRF/doc: fix 32-bit build failure on doc example

sizeof(size_t) does not always equal sizeof(uint64_t)

* host: docker: remove centos-7-i386 build

it doesn't work, and this is an edgy-enough case that i think we'll be
okay with the ubuntu i386 builds

* host: libbladeRF/bladerf2: explicitly note constant types

On non-64-bit platforms, these integer constants can be ambiguous, which
throws compiler warnings.

* host: libbladeRF_test: explicitly note constant types

On non-64-bit platforms, these integer constants can be ambiguous, which
throws compiler warnings.