host: libbladeRF: detect possible bladeRF2 during bladeRF1 init (#572)
If a pre-2.2.0 firmware is flashed to a bladeRF2 device, there is no way
to fix the problem without jumpering into bootloader because the
initialization aborts with a BLADERF_ERR_UNEXPECTED.

Here, we're looking for the bladeRF2 FPGA signatures, and if we find
them, we return from bladerf1_open() early after printing some useful
tips. This allows firmware flashing to occur, but not much else because
the FPGA is not loaded.