libbladeRF: adjust list of RX gain modes (#601)
On the bladeRF 1, remove the redundant "automatic" setting
and rename "default" to "automatic". Also directly use
enum values instead of the reverse-compat #defines.

On the bladeRF 2, rename "default" to "automatic".

Intent: make it more clear whether the gain mode is automatic
or manual, without having to "know" that the default is automatic.

Fixes #601