libbladeRF: implement bladerf1_{get,set}_rf_port[s]
This implements the RF port getters/setters for the bladeRF 1.

It is currently only implemented for the RX side, as the
low-level functions to handle the TX side haven't been written.
bladerf_get_rf_port on a bladeRF 1 TX channel will specify
"auto" as the port name, and bladerf_set_rf_port will just
return 0.