bladeRF-cli: ensure device is initialized after deferred open
If the --load-fpga, --flash-firmware, or --flash-fpga options
are present on the command line, we temporarily set the
BLADERF_FORCE_NO_FPGA_PRESENT environment variable when
opening the device, to trick bladerf_open into not performing
a device initialization before we have the hardware fully
ready to go.

The --load-fpga calls bladerf_load_fpga, which ultimately
calls the correct bladerf*_initialization function to complete
the initialization, so this is fine.

However, the --flash-* options do *not* do this, and therefore,
flashing firmware or FPGA leaves the device in a half-initialized
state (specifically STATE_FIRMWARE_LOADED). If the FPGA is
in the unconfigured state, this is an accurate and consistent,
if useless, state.

But, if the FPGA *has* been configured by a previous bladerf_open
or auto-loading from flash, we have two conflicting indicators:
dev->board_data->state indicates that the FPGA is NOT loaded, but
bladerf_is_fpga_configured() (which queries the FX3 for the state
of the FPGA's CONF_DONE pin) indicates that the FPGA *is* loaded.

The upshot of this: at the end of main(), bladeRF-cli calls
cli_state_destroy(), which calls rxtx_shutdown(), which calls
disarm_triggers(). If bladerf_is_fpga_configured() is true,
then disarm_triggers() calls bladerf_trigger_init(), which
quickly trips out on the CHECK_BOARD_STATE guard.

The fix for this is to add a bool to the rc_config struct, to
remember that we need to close and re-open the device normally
after the FPGA/firmware operations are completed.

This commit also clang-formats the modified functions, cleans
up some of the printfs, and improves the wording of the
informational messages during these operations.

Fixes #612