hdl/bladerf1: move U_agc_lms to 80 MHz clock domain
Move the bladerf_agc_lms_drv component into the 80 MHz clock domain.

This ensures that it has a consistent clock to work with, instead of the
variable rx_clock. This ensures the SPI clock will always be 20 MHz, and
that everything is passing between known clock domains.

Adds the requisite synchronizers, etc.