top: CMakeLists: if enabled, always build fx3
CMake assumes that external projects are not modified. However,
this is not always the case. (e.g. if you're modifying both fw
and host simultaneously and want to use your IDE's build system
interface)

There do not seem to be any side effects from this, and though
we generally recommend building host/ and fx3_firmware/
separately, this is still a useful capability to have.