host: common/devcfg: check board name before setting gains
Wrap bladeRF1-only gain functions around a conditional so they are not
executed on bladeRF2 boards.

This commit is intended to be a stopgap to get the quick tune tests to
build and function on bladeRF2 devices. A follow-up commit using the
more generic 'system gain' functions would be optimal.