nios: add fast lock functions for bladeRF2, increase RAM
Adds fast lock functions to save/load/recall, and change port/spdt
settings:

 - ad9361_fastlock_save(): save fast lock profile from AD9361 to BBP
 - ad9361_fastlock_load(): load fast lock profile from BBP to AD9361
 - ad9361_fastlock_recall(): use a previously loaded fast lock profile
 - ad9361_rfport_select(): change RF port selection on AD9361
 - ad9361_rfspdt_select(): change external RF SPDT switch settings

Increased the RAM size of the Nios from 16 KB to 32 KB in order to
hold 512 AD9361 fast lock profile structures (256 RX profiles, 256 TX
profiles). Each profile contains 20 bytes, of which 16 bytes are
consumed by the AD9361 profile data.

The value of 256 RX/TX fast lock profiles was chosen because it seemed
high enough for most applications, but not so large that it would waste
a lot of FPGA resources. However, the number of profiles may be
increased or decreased (as resources allow) to meet application demand.