hdl: nios: devices_rfic: add standby state, refactor initialization
hdl: nios: devices_rfic: add a standby initialization state

this will be used to indicate that the board is not actively
being driven, but is still initialized and ready to go

improve debugging output

enforce a hard reset at the start of initialization

hdl: nios: devices_rfic: perform init/tuneup on FPGA

Move some of the initialization tasks, previously done on the host and
forwarded up, to the Nios.

Also wee bit of refactoring/cleanup.