host: bladerf2: reset RFIC on close after doing retunes
For expedience, quick retunes directly control the RFIC via SPI.
This works fine as it is, but it puts the RFIC into an unexpected
state, and a reset is needed.

This commit fully de-initializes the RFIC during bladerf_close, if
a bladerf_get_quick_tune is performed sometime during the session.