host: udev: split up rules files by hardware platform
The underlying driver for this: Debian/Ubuntu's libbladerf1
package installs the udev rules itself, instead of having a
separate libbladerf-udev package like our PPA does. So,
having libbladerf1 installed creates a conflict with installing
libbladerf2, which is not good.

End result is that we either need to conflict libbladerf1, or
put the udev rules from libbladerf-udev somewhere else, and
there are fewer side effects with the latter option.