libbladerf: correctly report num_samples for SC8_Q7 in async api
The async API has two codepaths for reporting samples to a user-provided
callback depending on whether packet metadata is requested. In the path
where it is requested, the code correctly uses bytes_to_samples() to
report num_samples. In the path where no metadata is requested, the code
uses the previously hardcoded call to bytes_to_sc16q11(), which
underreports num_samples by half for SC8_Q7 format samples. This appears
to have been an oversight, and this patch corrects the non-metadata path
to match the metadata path.