libbladeRF: gain_calibration: added gain cal auto loading support
If the the file location is not specified, using a NULL file location,
then try loading the expected default file name from one of the
find_file locations.

The test renames the files to the default serial based file names and
stores them alongside the binary (one of the find_file locations).