libbladeRF: Return error code instead of asserting in version_check
Specifically, when there's a missing compat table entry. All callers
specify that they want to be forgiving of version mismatches, so this
seems in the spirit of the call tree.

Ran into this error when playing with a custom-loaded FPGA; if I had
flashed the FPGA instead of simply loading it, this could have bricked
the blade (or at least forced me into a very involved recovery process).